Fix typo in "Predicting New Observations" section
@bob-carpenter I think "at bats" should be "hits" in this sentence

"Rather than plotting the number of at bats on the vertical axis, we have standardized all the predictions and outcomes to a success rate"

because hits is the outcome being predicted. This didn't seem worthy of a pull request so I'm just pushing the change, but if I'm wrong feel free to revert of course.